* Added support for context menus and object property lists, which have been broken since the new networking was implemented.

+++ trunk/UltimaXNA/UltimaXNA/Network/Packets/Client/RequestContextMenuPacket.cs
@@ -10,8 +10,8 @@ namespace UltimaXNA.Network.Packets.Client
         public RequestContextMenuPacket(Serial serial)
             : base(0xBF, "Context Menu Request")
         {
-            Stream.Write((short)0x13);
-            Stream.Write(serial);
+            Stream.Write((short)0x13); // subcommand 0x13, request context menu
+            Stream.Write((int)serial);
         }
     }
 }
